+++ R/truncated.R
@@ -9,6 +9,32 @@
 #' @param lower,upper The range of values to keep from a distribution.
 #'
 #' @name dist_truncated
+#'
+#' @examples
+#' dist <- dist_truncated(dist_normal(2,1), lower = 0)
+#'
+#' dist
+#' mean(dist)
+#' variance(dist)
+#'
+#' generate(dist, 10)
+#'
+#' density(dist, 2)
+#' density(dist, 2, log = TRUE)
+#'
+#' cdf(dist, 4)
+#'
+#' quantile(dist, 0.7)
+#'
+#' if(requireNamespace("ggdist")) {
+#' library(ggplot2)
+#' ggplot() +
+#'   ggdist::stat_dist_halfeye(
+#'     aes(y = c("Normal", "Truncated"),
+#'         dist = c(dist_normal(2,1), dist_truncated(dist_normal(2,1), lower = 0)))
+#'   )
+#' }
+#'
 #' @export
 dist_truncated <- function(dist, lower = -Inf, upper = Inf){
   vec_is(dist, new_dist())