+++ NEWS.md
@@ -4,6 +4,7 @@
 
 * Added `dist_missing()` for representing unknown or missing (NA) distributions.
 * Documentation improvements.
+* Added `cdf()` method for `dist_sample()` which uses the emperical cdf.
 
 ## Bug fixes